fix: eliminate C4244 type conversion warnings in test code and topk (#4757)
MSVC generates C4244 warnings when implicit type conversions lose precision
(e.g., double→float, size_t→int). Fixed 136 warnings across 4 files:

1. **eigen_memory_tree_test.cc** (41 fixes)
   - Changed `std::make_pair(int, int/double)` to use proper types
   - Now uses `uint64_t{n}` for indices and `f` suffix for floats
   - Matches `emt_feats = std::vector<std::pair<uint64_t, float>>`

2. **igl_simulator.h** (6 fixes)
   - Changed integer/double literals to float literals with `f` suffix
   - Fixed `std::map<std::string, float>` initializers

3. **simulator.cc** (1 fix)
   - Added explicit cast from `size_t` to `int` in return statement
   - Function returns `std::pair<int, float>` but used `size_t` index

4. **topk.cc** (1 fix)
   - Changed `std::accumulate` initial value from `0` to `size_t{0}`
   - Prevents implicit conversion during accumulation

Eliminates all C4244 warnings:
- Windows 2022: 126 warnings (63 Debug + 63 Release)
- Windows 2019: 10 warnings (5 Debug + 5 Release)

+++ vowpalwabbit/core/tests/simulator.cc
@@ -198,7 +198,7 @@ std::pair<int, float> cb_sim::sample_custom_pmf(std::vector<float>& pmf)
   for (size_t index = 0; index < pmf.size(); ++index)
   {
     sum_prob += pmf[index];
-    if (sum_prob > draw) { return std::make_pair(index, pmf[index]); }
+    if (sum_prob > draw) { return std::make_pair(static_cast<int>(index), pmf[index]); }
   }
   THROW("Error: No prob selected");
 }
